Benchmark large-scale random number generation

This function focuses on benchmarking the performance of random number generators when generating large quantities of random numbers. It is particularly useful for evaluating the scaling behavior and memory efficiency of different generators.

Usage

benchmark_large_generation(
  n_values = c(1e6, 5e6, 1e7, 5e7),
  configs = list(
    "default" = list(),
    "optimized" = list(buffer_size = 1e6)
  ),
  generators = list(
    "qiprng" = function(n) generatePRNG(n),
    "base_r" = function(n) runif(n)
  ),
  repetitions = 3
)

Arguments

n_values

Vector of large sample sizes to benchmark

configs

List of QIPRNG configurations to test

generators

List of generator functions to compare

repetitions

Number of repetitions for each benchmark

Value

A list containing:

results: Raw benchmark data

summary: Summary statistics for each generator and configuration

plots: A list of ggplot objects visualizing the results

See also

Examples

if (FALSE) { # \dontrun{
# Compare QIPRNG to base R for large sample sizes
results <- benchmark_large_generation(
  n_values = c(1e6, 5e6),
  configs = list(
    "default" = list(),
    "large_buffer" = list(buffer_size = 100000)
  ),
  repetitions = 2
)

# View the results
print(results$summary)
print(results$plots$time_plot)
} # }