Profile different QIPRNG configurations

This function profiles the performance of different QIPRNG configurations to help users identify the optimal settings for their specific use case.

Usage

profile_qiprng_config(
  config_list = list(
    "default" = list(),
    "crypto" = list(use_crypto_mixing = TRUE),
    "high_precision" = list(mpfr_precision = 128),
    "normal" = list(distribution = "normal")
  ),
  n = 10000,
  repetitions = 10
)

Arguments

config_list

List of named configurations to test

n

Number of random numbers to generate for each test

repetitions

Number of benchmark repetitions

Value

A data frame with benchmark results for each configuration, including:

config: Configuration name

mean_time: Mean execution time in milliseconds

median_time: Median execution time in milliseconds

relative_speed: Performance relative to the default configuration

See also

Examples

if (FALSE) { # \dontrun{
# Profile several QIPRNG configurations
profiles <- profile_qiprng_config(
  config_list = list(
    "default" = list(),
    "crypto" = list(use_crypto_mixing = TRUE),
    "high_precision" = list(mpfr_precision = 128),
    "large_buffer" = list(buffer_size = 10000)
  ),
  n = 5000,
  repetitions = 3
)

# View the results
print(profiles)
} # }